+++ client/ayon_core/lib/__init__.py
@@ -113,7 +113,6 @@ from .transcoding import (
     convert_ffprobe_fps_to_float,
     get_rescaled_command_arguments,
     get_media_mime_type,
-    get_default_reviewable_layers
 )
 
 from .plugin_tools import (
@@ -220,7 +219,6 @@ __all__ = [
     "convert_ffprobe_fps_to_float",
     "get_rescaled_command_arguments",
     "get_media_mime_type",
-    "get_default_reviewable_layers",
 
     "compile_list_of_regexes",
