+++ client/ayon_core/pipeline/publish/lib.py
@@ -1413,3 +1413,15 @@ def _get_last_version_files(
     ]
 
     return (version_entity, repre_file_paths)
+
+
+def get_default_reviewable_layers(project_settings: dict) -> list[str]:
+    """Get default reviewable layers from project settings.
+
+    Args:
+        project_settings (dict): Project settings.
+
+    Returns:
+        list[str]: List of default reviewable layers.
+    """
+    return project_settings["core"]["reviewable_layers"]["review_layers"]
